requirements.txt
# 深度学习框架
# 注意：建议使用官方命令安装以匹配你的 CUDA 版本 (例如 CUDA 12.1)
# pip install torch --index-url https://download.pytorch.org/whl/cu121
torch>=2.4.0

# 图神经网络库
# 注意：DGL 必须与 PyTorch 和 CUDA 版本匹配
# pip install dgl -f https://data.dgl.ai/wheels/cu121/repo.html
dgl>=2.4.0
ogb>=1.3.6

# 科学计算与数据处理
numpy>=1.26.0
scipy>=1.10.0
pandas>=2.0.0
scikit-learn>=1.5.0

# 配置文件解析
pyyaml

# 进度条工具
tqdm

# 编译工具 (加速 C++ 扩展编译)
ninja